Перенос отправки данных в api
