+++ js/api.js
@@ -10,6 +10,30 @@ const getData = () => {
     });
 };
 
+const sendData = (onSuccess, onError) => {
+
+  const submitFormElement = document.querySelector('.ad-form');
+  submitFormElement.addEventListener('submit', (evt) => {
+    evt.preventDefault();
+
+    const formData = new FormData(evt.target);
+
+    fetch('https://24.javascript.pages.academy/keksobooking', {
+      method: 'POST',
+      body: formData,
+    }).then((response) => {
+      if (response.ok) {
+        onSuccess();
+      } else {
+        onError();
+      }
+    }).catch(() => {
+      onError();
+    });
+  });
+};
+
 export {
   getData,
+  sendData
 };
